// src/app/components/ui/User/types/ListSelectFilter.enum.ts
export enum ListSelectFilter {
  title = 'Title',
  location = 'Location',
  department = 'Department',
  evaluator = 'Evaluator',
  strategy = 'Strategy',
  form = 'Form',
  status = 'Status',
  LOC = 'LOC',
  selfAssessment = 'Self_Assessment',
  missingEvaluations = 'Missing_Evaluations',
  perfSummary = 'Performance_Summary',
}